Datamap and Prettymap
Made it so that there's one datamap that is just basic colors for the nations which is used to figure out when the mouse hover over them, and a pretty map which is for visuals. And also I made a nation-object that has that nation's prettymap-image that gets drawn on top of prettymap when you hover over it.