Work to fix bugs from parsing rework
Did a lot of work to fix bugs that resulted from the rework of the parsing system, mainly those in the database-storage-thingies. It's not done yet though.

As part of the above I introduced a new system for parsing users, so that we don't just parse the current game mods. It seems to work in itself.

Also gitignore refuses to ignore .txt-files for some reason, so whatever then, let's push them I guess!